INQ-46 INQ-48 INQ-49 Added Public Data Portal (First Cut)
This commit adds a basic data portal that allows the public to
search published data sets (repositories and Data nodes). There is
a "live" search feature that displays data as well as a detail view
that allows users to browse and export data from a specific repository.

This could use more UI styling and needs to add some basic browse
features as well (but I'm not sure we've got the metadata to support
them).