INQ-47 Support Repo Browsing in the Portal
This adds the ability to browse by a few basic criteria in the
portal. To do so I've added a few more Vuex actions and a new
endpoint for portal browsing.

Additionally, to support the "featured" option, I've added an
admin-only screen that lets them see all of the repositories in the
database. This should be improved, but at least it gives us the
option now/something to build on. It's only function right now is
to toggle the featured repos.