INQ-17 Reorganized Upload Result Screen
Improved Upload result screen to give an easier to read experience
for users and a more logical layout of results. Errors no longer
take up so much space and do not appear at all if they do not occur.

This also gives more space to the chart, and allows for further charts
to be added easily (which will be worked on soon)