INQ-35 Added popups and other tips to upload process
This comprises the initial batch of tooltips for Inquisite, and
includes some supporting code enabling them.

It was suggested that a custom component be created but instead
using the standard bootstrap-vue compenets make this easier. This
commit integrates that library into the ui and allows us to start
using them as drop-ins. A lot of the interface could eventually
be re-written to use this library, but for the time being I will
just be using it to replace functionality that otherwise would
have to be implemented in jQuery